requirements.txt
Django==4.0.2
django-mptt==0.13.4
django-tables2==2.4.1
django-filter==21.1
django-bootstrap3==21.2
Faker==12.3.0
psycopg2-binary==2.9.3